requirements.txt
Django>=4.2
protobuf
google-genai
